int in float geändert

+++ 2O2P/src/oo2cad/shapes/MasterShape.java
@@ -9,34 +9,35 @@ package oo2cad.shapes;
  */
 public class MasterShape {
 
-	private int x;
-	private int y;
-	private int height;
-	private int width;
+	private float x;
+	private float y;
+	private float height;
+	private float width;
 	
-	public int getX() {
+	public float getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
-	public int getHeight() {
+	public float getHeight() {
 		return height;
 	}
-	public void setHeight(int height) {
+	public void setHeight(float height) {
 		this.height = height;
 	}
-	public int getWidth() {
+	public float getWidth() {
 		return width;
 	}
-	public void setWidth(int width) {
+	public void setWidth(float width) {
 		this.width = width;
 	}
+	
 	
 }
